requirements.txt
Flask==2.3.3
requests==2.31.0
gspread==6.1.3
oauth2client==4.1.3
PyYAML==6.0.1
